Changes log level of some messages.
As requested in pull request from bdr99:
Please change lines 175, 179, 187, and 259 to info. Those log messages are completely normal and expected during regular usage of this library, so they should not be warnings.
Please change lines 170 and 236 back to debug. I think the only reason someone would ever need to view the raw request and response payloads is if they were working on developing this library.